requirements.txt
Django~=3.2.3
django-registration==3.1.2
Pillow==8.2.0
PyQRCode==1.2.1
pypng==0.0.20
fpdf2==2.3.5
easy-thumbnails==2.7.1
gunicorn
psycopg2-binary
django-crontab==0.7.1
python-dateutil==2.8.1
glQiwiApi==1.0.0

aiogram~=2.12
environs~=8.0.0
aiohttp~=3.7.4.post0
requests~=2.25.1

pyTelegramBotAPI==3.7.9